(gst_plugins) fix issues of klockwork static code analysis result
Issues in gstreamer_plugin_based_on_OpenVINO
1. Freeing freed memory due to missing copy constructor - 8/9
2. C/C++ Warnings - 2/10 is in this project, 8/10 in 3rd header files.
3. Memory Leaks - 1/1
4. Null Pointer Dereference - 8/8
5. Mismatched Return Types - 0/12, all are in 3rd header file.
6. Unreachable Code - 2/2
7. Use of Uninitialized Data - 2/4, other is in 3rd header file.

summary: 46 issues
       fixed: 23
       3rd header files: 23

+++ gstreamer_plugin_openVINO/gst-libs/algo/algobase.cpp
@@ -93,7 +93,7 @@ static void try_process_algo_data(CvdlAlgoData *algoData)
                       (uint32_t)objectData.rectROI.width,
                       (uint32_t)objectData.rectROI.height};
 
-    if(crop.width<=0 || crop.height<=0 || crop.x<0 || crop.y<0) {
+    if((int)crop.width<=0 || (int)crop.height<=0 || (int)crop.x<0 || (int)crop.y<0) {
         GST_ERROR("Invalid  crop = (%d,%d) %dx%d", crop.x, crop.y, crop.width, crop.height);
         objectData.flags |= CVDL_OBJECT_FLAG_DONE;
         try_process_algo_data(algoData);
